requirements.txt
python-telegram-bot==20.0
python-telegram-bot[rate-limiter]
openai==0.27.0
requests==2.31.0